+++ config.js
@@ -193,15 +193,15 @@ var cfg = {
             exclude: /(node_modules|bower_components|vendor)/
           })
           loaders.push({
-            test: /\.css/,
-            loader: ExtractTextPlugin.extract('css-loader?sourceMap'),
-            exclude: /(node_modules|bower_components|vendor)/
-          })
-          loaders.push({
-            test: /\.less/,
+            test: /\.less$/,
             loader: ExtractTextPlugin.extract('css-loader?sourceMap!less-loader?sourceMap'),
             exclude: /(node_modules|bower_components|vendor)/
           })
+          // loaders.push({
+          //   test: /\.css$/,
+          //   loader: ExtractTextPlugin.extract('style-loader!css-loader?sourceMap'),
+          //   exclude: /(node_modules|bower_components|vendor)/
+          // })
         }
 
         return loaders