+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 [Diff](https://github.com/kvz/lanyon/compare/v0.0.33...v0.0.34)
 
 - [x] Fix bad Git ignores
+- [x] Revert plain css files to extract text plugin
 
 # v0.0.33 (bad build)